+++ VK9-Library/Utilities.cpp
@@ -45,7 +45,7 @@ VkShaderModule LoadShaderFromFile(VkDevice device, const char *filename)
 	if (fp != nullptr)
 	{
 		fseek(fp, 0L, SEEK_END);
-		size_t dataSize = ftell(fp);	
+		int32_t dataSize = ftell(fp);	
 		fseek(fp, 0L, SEEK_SET);
 		void* data = (uint32_t*)malloc(dataSize);
 		if (data != nullptr && fread(data, dataSize, 1, fp))
@@ -103,7 +103,7 @@ VkShaderModule LoadShaderFromResource(VkDevice device, WORD resource)
 			HGLOBAL hData = LoadResource(dllModule, hRes);
 			if (NULL != hData)
 			{
-				size_t dataSize = SizeofResource(dllModule, hRes);
+				int32_t dataSize = SizeofResource(dllModule, hRes);
 				uint32_t* data = (uint32_t*)LockResource(hData);
 
 				moduleCreateInfo.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
